Making a common file utils for all the common functions used Removing the logics out of the view file and fixing a small path issue while querying the local peer adding back wikiwoods for indexing
